Fix bugs, add Gradle wrapper, update SDK target
- Fix getString() called on Dispatchers.Default (must be on Main thread)
- Remove redundant toInt().toFloat() casts in FaceSwapEngine
- Rename project from "Android Face Swap App" to "AndroidFaceSwapApp"
  (spaces in project names can cause Gradle issues)
- Bump compileSdk/targetSdk 34 → 35 (Android 15)
- Add generated gradlew / gradlew.bat / gradle-wrapper.jar so the app
  builds out-of-the-box with `./gradlew assembleDebug`
- Update .gitignore to track wrapper jar and exclude .gradle/ cache

+++ android-face-swap-app/app/src/main/java/com/example/faceswap/FaceSwapEngine.kt
@@ -177,8 +177,8 @@ object FaceSwapEngine {
         canvas.save()
         canvas.clipPath(path)
         canvas.drawRect(
-            (box.left  - r).toInt().toFloat(), (box.top  - r).toInt().toFloat(),
-            (box.right + r).toInt().toFloat(), (box.bottom + r).toInt().toFloat(),
+            box.left  - r, box.top  - r,
+            box.right + r, box.bottom + r,
             Paint(Paint.ANTI_ALIAS_FLAG).apply {
                 shader = gradient
                 xfermode = PorterDuffXfermode(PorterDuff.Mode.SRC_ATOP)
